create package.json
creates a package.json file for defence-labs.live. this package.json contains all the necessary information for a successful deployment of the defence-labs.live project.

the package.json contains the name, version, description, main, scripts, repository, author, license, bugs and homepage of the project. all of this information is necessary for the successful deployment of the project.