adding myapp component and home page
adds a new component, `myapp`, to the project as well as a new page, `home`.

the new `myapp` component is located in `pages/_app.js` and is responsible for providing platform-level components to the application. the `home` page is located in `pages/index.js` and is responsible for rendering the "hello world" message on the home page.